Fix TypeError: Cannot read property 'toggleActivated' of null
Hi,

Thanks for your work on this neat plugin. 

Set mouse 'click' option to TREE_ACTIONS.TOGGLE_EXPANDED, then click on first node in the tree, then hit "space" bar on your keyboard. An error is raised. Proposed fix may not be the correct way to fix that, I guess node should be checked before being sent to  this line where toggleActivated is called on it.

Best regards,

Greg