Set treeDraggedElement to null on drop
TreeDraggedElement is set on dragstart in tree-drag.directive. That directive also unset the element on dragend. However, this is only done if the position for the dragend is on the same element as the dragstart. The element should be removed also when the dragging is stopped somewhere else. This fix sets the element to null in the onDrop method in tree-drop.directive.